Resolve race condition by guaranteeing that onDestroyed is first on the right queue
The HalDevManager destroys interfaces and then posts the onDestroyed to
the Handler. However, if the HAL blocks or semi-blocks the interface
destruction then it is possible that we'll get notification from the HAL
about the interfaces disapearing before the onDestroyed cleans up the
data structures.

Guarantee that the onDestroyed is called before the HAL callback.

1. Make sure to post the onDestroyed to front of queue if both are on
   the same thread.
2. Make sure to pre-execute the onDestroyed callback if both are on
   different threads.

Add debugging messages.

Bug: 211919010
Bug: 198124924
Bug: 201255954
Bug: 202905932
Bug: 209564054
Test: atest android.net.wifi
Test: atest com.android.server.wifi
CRs-Fixed: 3113789
Change-Id: I5a2b96ad98e4feb3802ae82198021860c4915738
(cherry-picked from 4d163bdd52412b8c13053f5bae7561a8159798c1)