wifi: Fix exception when configuring macdomized MAC address
The calculate MAC address might return null which causes exception.
Change the design to store persistent randomized MAC address in
configuration file and use it when enabling Soft AP and use a
randomized MAC when issue happen.

So the current design is

First to use getPersistentRandomizedMacAddress() when it is NOT
null. And save persistent randomized MAC address when
a. generate default tethered configuration
b. generate lohs configuration
c. setApConfiguration is called (i.e. The SSID might be updated)

And generate a randomized MAC again when issue happened.

So for tethered hotspot use case when issue happened
The persistent randomized MAC address would be randomized MAC but
it is stored in SoftApConfiguration. So frameworks use the randomized
MAC address same as the API: getPersistentRandomizedMacAddress.

For lohs use case when issue happened. There are two cases
1. No specific lohs configuration. i.e. use default lohs config.
(The SSID would change every time). So the persistent randomized
MAC address is randomized MAC again which would NOT break the use
case.

2. Start lohs with customized configuration when issue happen.
a. If #getPersistentRandomizedMacAddress() is null,
it means the caller generate this configuration by itself, i.e.
Not frameworks provides it. So it doesn't matter if frameworks
use a randomized MAC since caller doesn't care about it.

b. If SoftApConfiguration#getPersistentRandomizedMacAddress() is NOT
null. Frameworks still use this getPersistentRandomizedMacAddress to
enable lohs even if it is generated from ramdonization, not a
calculated result.

Bug: 225287753
BUg: 226627115
Test: atest -c FrameworksWifiTests
CRs-Fixed: 3152829
Change-Id: I3d8fc65728fbbb8ca78b7fe71b66d7db0fe712c7
(cherry picked from commit 13b88a80718d44c43dadbb62e95439bef7744718)