SoftAp: Schedule timeout for SoftAp in OWE and OWE transition mode
CMD_AP_INFO_CHANGED is not triggered by Hostapd Vendor Hal when SoftAp
is started in OWE and OWE transition modes. Due to this, the idle shutdown timer
is not started and the SoftAp would not be turned off after the timeout
when no devices are connected.

Fix this by scheduling timeout while SoftAp starts in OWE transition
mode and OWE mode from Hostapd vendor hal.

CRs-Fixed: 3201152
Change-Id: I3b32b710483ce78335f9cc0bf00bd35fa54b429f