HDM: update Interface combination based on driver capabilities.
Inteface combination is a build time configuration which is defined
for each target. For the targets where host driver with different
interface combination can be used, build time configuration breaks
the functionality.

Now that Wifi Hal supports query of iface combination, reload the
driver supported combination once first interface is created.
Since this combination is not going to change, it is stored to
WifiConfigStore.xml for static lookup when driver is not loaded.
Thus, maintains the backward compatibility.

Change-Id: Ic24332cd0b3c22380e9b1fc8ee7af06b5fae6cac
CRs-Fixed: 3230634