Disallow concurrent STA to connect affiliated BSSIDs
Consider a STA + STA setup is  connecting to an MLO AP
'{BSSID_2_4, BSSID_5} MLD MAC Address', where BSSID_2_4 and BSSID_5 are
affiliated link BSSIDs. SSID is same for all the APs.

	Case 1: Legacy Station + Legacy Station
	---------------------------------------

	a) legacy_sta1 connects to BSSID_2_4.
	b) legacy_sta2 is not allowed to connect BSSID_2_4 (current
	   restriction).
	c) legacy_sta2 connects to BSSID_5.

	Case 2: MLO Station + Legacy Station
	------------------------------------

	a) mlo_sta1 connects to 'MLD MAC Address {BSSID_2_4, BSSID_5}'.
	b) legacy_sta2 is not allowed to connect either BSSID_2_4
	   or BSSID_5. (extending current restiction)

	case 3: MLO Station + MLO Station
	---------------------------------

	a) mlo_sta1 connects to 'MLD MAC Address {BSSID_2_4, BSSID_5}'
	b) mlo_sta2 is not allowed to connect 'MLD MAC Address
	   {BSSID_2_4, BSSID_5}' (extending current restriction)

In summary, if one STA is MLO, then the other STA (whether MLO or
legacy), can not use any of the BSSIDs of the MLO AP. If we are
connected in legacy mode, then stick to the current restrictions.

Bug: 240186395
Test: atest com.android.server.wifi
CRs-Fixed: 3313211
Change-Id: I2050361a0812207b9b7d62616c4f14210179426e
(cherry-picked from fa3f216be6ff61eec95cd9e18a180857ce8421a2)